refactor(NodeService): move evaluateTransitionLogic method to StudentNodeService. It's only used by students.
